Extends next to declare a winner when entries.size === 1
'next' now has an if statement looking for the entries.size.
When this === 1 it removes the vote node, entries node, and sets the
winner node to the only entry in 'entries'.
It removes the other nodes instead of just returning a new map with
only the winner node to avoid potential issues in the future with
destroying unrelated data. (Not returning it)