Integrates the socket.io server with Redux
By passing the redux store into the socket.io start server function
the server now has the ability to send the state to people connected.
Adds an on connect event listener set to send the latest state
Adds a function that sends a JSON state snapshot to everyone
connected anytime the state change.